config 클래스 내 client_max_body_size 타입 int -> long long으로 수정

+++ src/config/model/Config.cpp
@@ -1,6 +1,11 @@
 #include "Config.hpp"
 
-Config::Config() : _auto_index(false), _listen(-1), _client_max_body_size(1024 * 1024), _server_name("_"), _index("index.html") {}
+Config::Config()
+	: _auto_index(false),
+	  _listen(-1),
+	  _client_max_body_size(1024 * 1024),
+	  _server_name("_"),
+	  _index("index.html") {}
 
 Config::Config(const Config& other) { *this = other; }
 
@@ -19,7 +24,7 @@ bool Config::getAutoIndex() const { return _auto_index; }
 
 int Config::getListen() const { return _listen; }
 
-int Config::getClientMaxBodySize() const { return _client_max_body_size; }
+long long Config::getClientMaxBodySize() const { return _client_max_body_size; }
 
 const std::string& Config::getServerName() const { return _server_name; }
 
@@ -35,7 +40,7 @@ void Config::setAutoIndex(bool auto_index) { _auto_index = auto_index; }
 
 void Config::setListen(int listen) { _listen = listen; }
 
-void Config::setClientMaxBodySize(int client_max_body_size) {
+void Config::setClientMaxBodySize(long long client_max_body_size) {
 	_client_max_body_size = client_max_body_size;
 }
